+++ levior/filters/__init__.py
@@ -9,6 +9,7 @@ from trimgmi import LineType
 from trimgmi import Line
 
 from .text import rm_bracketed_digits  # noqa
+from .text import sub_bracketed_digits  # noqa
 from .text import text_filter  # noqa
 from .text import get_out  # noqa
 from .text import uppercased  # noqa